refactor: components, routes 컴포넌트 파일 이름 -> index.tsx로 변경
