--- src/routes/Weathers/Cities/Cities.module.scss
@@ -1,145 +1,0 @@
-@use '/src/styles/constants/colors';
-@use '/src/styles/constants/levels';
-@use '/src/styles/mixins/flexbox';
-@use '/src/styles/mixins/responsive';
-@use '/src/styles/mixins/position';
-
-.city {
-  h1 {
-    margin-top: 14px;
-    font-size: 24px;
-    font-weight: 600;
-    color: colors.$WHITE;
-    text-align: center;
-  }
-
-  form {
-    display: flex;
-    margin-top: 14px;
-
-    .leftWing {
-      flex: 1;
-      margin-right: 8px;
-    }
-
-    input {
-      width: 100%;
-      height: 44px;
-      padding: 0 14px;
-      color: colors.$WHITE;
-      background-color: rgba(colors.$WHITE, 15%);
-      border-radius: 12px;
-
-      + input {
-        margin-top: 8px;
-      }
-
-      &::placeholder {
-        color: colors.$GRAYA;
-      }
-    }
-
-    button {
-      flex-shrink: 0;
-      width: 60px;
-      color: colors.$WHITE;
-      background-color: rgba(colors.$WHITE, 15%);
-      border-radius: 12px;
-    }
-  }
-
-  .forecast {
-    display: flex;
-    flex-direction: column;
-    margin-top: 40px;
-
-    .placeholderMsg {
-      margin-top: 100px;
-      text-align: center;
-    }
-
-    h2 {
-      font-size: 14px;
-      font-weight: 600;
-    }
-
-    ul {
-      margin-top: 24px;
-    }
-
-    li {
-      padding: 24px 14px;
-      background-color: rgba(colors.$BLACK, 10%);
-      border-radius: 20px;
-
-      + li {
-        margin-top: 24px;
-      }
-
-      dl {
-        display: flex;
-
-        > div {
-          flex: 1;
-        }
-      }
-
-      dt,
-      dd {
-        text-align: center;
-      }
-
-      dt {
-        display: none;
-      }
-
-      time {
-        display: block;
-
-        &:first-child {
-          font-weight: 600;
-        }
-
-        &:last-child {
-          margin-top: 4px;
-        }
-      }
-
-      .temp {
-        font-size: 32px;
-
-        sup {
-          font-size: 10px;
-          vertical-align: top;
-        }
-      }
-    }
-  }
-
-  @include responsive.after(SD) {
-    ul {
-      display: flex;
-      flex-wrap: wrap;
-    }
-
-    li {
-      width: 240px;
-      height: 120px;
-      margin: 8px;
-
-      + li {
-        margin: 8px !important;
-      }
-    }
-  }
-
-  @include responsive.after(MD) {
-    width: 100%;
-    background: burlywood;
-  }
-
-  @include responsive.after(LD) {
-    width: 100%;
-    background: cadetblue;
-  }
-}
