--- src/routes/searchDiseases/SearchList/searchList.module.scss
@@ -1,27 +1,0 @@
-@use '/src/styles/constants/colors';
-@use '/src/styles/constants/levels';
-@use '/src/styles/mixins/flexbox';
-@use '/src/styles/mixins/responsive';
-@use '/src/styles/mixins/position';
-
-.list {
-  width: 700px;
-  height: 100%;
-  background-color: white;
-  border-radius: 10px;
-  margin-top: 10px;
-
-  .listContent {
-    width: 700px;
-    height: 50px;
-    line-height: 50px;
-    padding-left: 50px;
-
-    &:hover {
-      background-color: rgba(colors.$GRAYD, 90%);
-    }
-  }
-  .isFocus {
-    background-color: rgba(colors.$GRAYD, 90%);
-  }
-}
