--- src/types/todo.d.ts
@@ -1,5 +1,0 @@
-export interface ITodoItem {
-  id: number
-  title: string
-  done: boolean
-}
